requirements.txt
crewai
crewai[tools]
pydantic
python-dotenv
fastapi